renamed files to use consistent kebab-case
