+++ package.json
@@ -83,6 +83,8 @@
       "!src/**/*.d.ts"
     ],
     "moduleNameMapper": {
+      "^@/(.*)\\.js$": "<rootDir>/src/$1",
+      "^@/(.*)$": "<rootDir>/src/$1",
       "^(\\.{1,2}/.*)\\.js$": "$1"
     },
     "transform": {
