+++ src/cli/commands/list.ts
@@ -1,6 +1,6 @@
 import * as YAML from 'yaml';
-import { WorkflowEngine } from '../../core/WorkflowEngine.js';
-import { ConfigDiscovery } from '../../core/ConfigDiscovery.js';
+import { WorkflowEngine } from '../../core/workflow-engine.js';
+import { ConfigDiscovery } from '../../core/config-discovery.js';
 import { Collection } from '../../core/types.js';
 
 interface ListOptions {
