+++ tests/helpers/file-system-helpers.ts
@@ -1,4 +1,4 @@
-import { MockSystemInterface } from '../mocks/MockSystemInterface.js';
+import { MockSystemInterface } from '../mocks/mock-system-interface.js';
 import { crawlDirectoryStructure } from '../../scripts/generate-mock-fs.js';
 import * as path from 'path';
