+++ tests/unit/config-discovery.test.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
-import { ConfigDiscovery } from '../../src/core/ConfigDiscovery.js';
-import { MockSystemInterface } from '../mocks/MockSystemInterface.js';
+import { ConfigDiscovery } from '../../src/core/config-discovery.js';
+import { MockSystemInterface } from '../mocks/mock-system-interface.js';
 
 describe('ConfigDiscovery', () => {
   let mockSystemInterface: MockSystemInterface;
