+++ tests/unit/workflow-engine-status.test.ts
@@ -1,9 +1,9 @@
 import * as path from 'path';
 import * as YAML from 'yaml';
-import { WorkflowEngine } from '../../src/core/WorkflowEngine.js';
-import { ConfigDiscovery } from '../../src/core/ConfigDiscovery.js';
-import { MockSystemInterface } from '../mocks/MockSystemInterface.js';
-import { createEnhancedMockFileSystem } from '../helpers/FileSystemHelpers.js';
+import { WorkflowEngine } from '../../src/core/workflow-engine.js';
+import { ConfigDiscovery } from '../../src/core/config-discovery.js';
+import { MockSystemInterface } from '../mocks/mock-system-interface.js';
+import { createEnhancedMockFileSystem } from '../helpers/file-system-helpers.js';
 
 // Mock dependencies
 jest.mock('path');
@@ -161,7 +161,7 @@ describe('WorkflowEngine - updateCollectionStatus', () => {
     );
 
     // Mock getCurrentISODate for deterministic testing
-    jest.doMock('../../src/shared/dateUtils.js', () => ({
+    jest.doMock('../../src/shared/date-utils.js', () => ({
       getCurrentISODate: jest.fn().mockReturnValue('2025-01-21T11:00:00.000Z'),
     }));
